+++ assembly-adapter/src/main/java/me/panpf/adapter/AssemblyExpandableAdapter.java
@@ -16,8 +16,6 @@
 
 package me.panpf.adapter;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseExpandableListAdapter;
@@ -26,6 +24,8 @@ import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import me.panpf.adapter.expandable.ExpandableItemActor;
 import me.panpf.adapter.expandable.ExpandableItemStorage;
 import me.panpf.adapter.more.MoreItemFactory;
@@ -34,7 +34,6 @@ import me.panpf.adapter.more.MoreItemHolder;
 /**
  * 通用组合式 {@link BaseExpandableListAdapter}，支持组合式多 item，支持头、尾巴以及加载更多
  */
-@SuppressWarnings({"unused", "WeakerAccess"})
 public class AssemblyExpandableAdapter extends BaseExpandableListAdapter implements AssemblyAdapter {
 
     @NonNull
@@ -122,6 +121,12 @@ public class AssemblyExpandableAdapter extends BaseExpandableListAdapter impleme
         return storage.addHeaderItem(itemFactory);
     }
 
+    @NonNull
+    @Override
+    public <DATA> ItemHolder<DATA> addHeaderItem(@NonNull ItemHolder<DATA> itemHolder) {
+        return storage.addHeaderItem(itemHolder);
+    }
+
     @Override
     public List<ItemHolder> getHeaderItemList() {
         return storage.getHeaderItemList();
@@ -153,6 +158,12 @@ public class AssemblyExpandableAdapter extends BaseExpandableListAdapter impleme
         return storage.addHeaderItem(itemFactory);
     }
 
+    @NonNull
+    @Override
+    public <DATA> ItemHolder<DATA> addFooterItem(@NonNull ItemHolder<DATA> itemHolder) {
+        return storage.addFooterItem(itemHolder);
+    }
+
     @Override
     public List<ItemHolder> getFooterItemList() {
         return storage.getFooterItemList();
@@ -181,6 +192,12 @@ public class AssemblyExpandableAdapter extends BaseExpandableListAdapter impleme
     @Override
     public <DATA> MoreItemHolder<DATA> setMoreItem(@NonNull MoreItemFactory<DATA> itemFactory) {
         return storage.setMoreItem(itemFactory);
+    }
+
+    @NonNull
+    @Override
+    public <DATA> MoreItemHolder<DATA> setMoreItem(@NonNull MoreItemHolder<DATA> itemHolder) {
+        return storage.setMoreItem(itemHolder);
     }
 
     @Nullable